created export controller, separated export summary to 2 methods
