+++ packages/api/src/period/routes.ts
@@ -1,6 +1,7 @@
 import { Router } from '@awaitjs/express';
 import * as core from './controllers/core';
 import * as assignment from './controllers/assignment';
+import * as exportController from './controllers/export';
 
 // Period-routes
 const periodRouter = Router();
@@ -16,8 +17,15 @@ const adminPeriodRouter = Router();
 adminPeriodRouter.postAsync('/create', core.create);
 adminPeriodRouter.patchAsync('/:periodId/update', core.update);
 adminPeriodRouter.patchAsync('/:periodId/close', core.close);
-adminPeriodRouter.getAsync('/:periodId/export', core.exportPraise);
-adminPeriodRouter.getAsync('/:periodId/exportSummary', core.exportSummary);
+adminPeriodRouter.getAsync('/:periodId/export', exportController.exportPraise);
+adminPeriodRouter.getAsync(
+  '/:periodId/exportSummary',
+  exportController.exportSummary
+);
+adminPeriodRouter.getAsync(
+  '/:periodId/customExport',
+  exportController.customExport
+);
 adminPeriodRouter.getAsync(
   '/:periodId/verifyQuantifierPoolSize',
   assignment.verifyQuantifierPoolSize
