+++ packages/api/src/period/utils/core.ts
@@ -1,5 +1,4 @@
 import some from 'lodash/some';
-import axios, { AxiosResponse } from 'axios';
 import { NotFoundError } from '@/error/errors';
 import { PraiseModel } from '@/praise/entities';
 import { calculateGiverReceiverCompositeScore } from '@/praise/utils/score';
@@ -7,7 +6,6 @@ import { periodsettingListTransformer } from '@/periodsettings/transformers';
 import { PeriodSettingsModel } from '@/periodsettings/entities';
 import { settingValue } from '@/shared/settings';
 import { isQuantificationCompleted } from '@/praise/utils/core';
-import { transform } from './jsonTransformer';
 import { PeriodModel } from '../entities';
 import {
   periodDetailsGiverReceiverListTransformer,
@@ -21,9 +19,6 @@ import {
   PeriodDetailsGiverReceiver,
   PeriodDateRange,
   PeriodStatusType,
-  PeriodDetailsGiverReceiverDto,
-  TransformerOperateItem,
-  TransformerMap,
 } from '../types';
 
 /**
@@ -299,63 +294,4 @@ export const isPeriodLatest = async (
   if (latestPeriods[0]._id.toString() === period._id.toString()) return true;
 
   return false;
-};
-
-export const getExportTransformer = async (
-  url: string
-): Promise<TransformerMap> => {
-  let response: AxiosResponse | undefined = undefined;
-  try {
-    response = await axios.get(url);
-  } catch (error) {
-    throw new Error('Could not fetch transformer');
-  }
-
-  // TODO add schema validation
-  if (response) {
-    const buff = Buffer.from(response.data.content, 'base64');
-    return JSON.parse(buff.toString('utf-8')) as TransformerMap;
-  }
-  throw new Error('Unknown error');
-};
-
-export const getSummarizedReceiverData = (
-  data: PeriodDetailsGiverReceiverDto[],
-  customExportContext: string,
-  csSupportPercentage: number,
-  transformer: TransformerMap
-): Object[] => {
-  const exportContext = JSON.parse(
-    customExportContext
-  ) as typeof transformer.context;
-
-  const totalPraiseScore = data
-    .map((item) => item.scoreRealized)
-    // eslint-disable-next-line @typescript-eslint/restrict-plus-operands
-    .reduce((prev, next) => prev + next);
-
-  const context = {
-    ...exportContext,
-    ...{
-      totalPraiseScore: totalPraiseScore,
-      csWalletAddress: 'Test ETH address',
-      csSupportPercentage: csSupportPercentage,
-    },
-  };
-
-  const map = {
-    item: transformer.map.item,
-    operate: transformer.map.operate.map(
-      (operateItem: TransformerOperateItem) => {
-        return {
-          run: operateItem.run,
-          on: operateItem.on,
-        };
-      }
-    ),
-    each: transformer.map.each,
-  };
-
-  const result = transform(data, map, context);
-  return result;
 };
